Add delete functionality in all + some changes + add task module

+++ resources/views/tasks/list.blade.php
@@ -1,6 +1,6 @@
 @extends('layout.app')
-@section('title', 'User List | TSP - Task Management System')
-@section('pageTitle', 'User List')
+@section('title', 'Task List | TSP - Task Management System')
+@section('pageTitle', 'Task List')
 
 @section('content')
 
@@ -11,26 +11,33 @@
                 <table id="users-datatable" class="table dt-responsive nowrap w-100">
                     <thead>
                         <tr>
-                            <th>Name</th>
-                            <th>Action</th>
+                            <th>Title</th>
+                            <th>Description</th>
+                            <th>Project</th>
+                            <th>Priority</th>
+                            <th>Status</th>
                         </tr>
                     </thead>
 
                     <tbody>
-                        @foreach ($users as $user)
+                        @foreach ($tasks as $task)
                             <tr>
-                                <td>{{ $user->name }}</td>
-                                <td>
-                                    {{-- <a class="btn btn-info" href="{{ route('roles.show', $role->id) }}">Show</a> --}}
+                                <td>{{ $task->title }}</td>
+                                <td>{{ $task->description }}</td>
+                                <td>{{ $task->project->name }}</td>
+                                <td>{{ config('constants.PRIORITY_LIST')[$task->priority] }}</td>
+                                <td>{{ config('constants.STATUS_LIST')[$task->status] }}</td>
+                                {{-- <td>
+                                    <a class="btn btn-info" href="{{ route('roles.show', $role->id) }}">Show</a>
                                     @can('update-roles')
                                         <a class="btn btn-primary" href="{{ route('users.edit', $user->id) }}">Edit</a>
                                     @endcan
-                                    {{-- @can('delete-roles')
+                                    @can('delete-roles')
                                         {!! Form::open(['method' => 'DELETE','route' => ['roles.destroy', $role->id],'style'=>'display:inline']) !!}
                                             {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
                                         {!! Form::close() !!}
-                                    @endcan --}}
-                                </td>
+                                    @endcan
+                                </td> --}}
                             </tr>
                         @endforeach
                     </tbody>
